+++ src/components/Input/example/index.js
@@ -5,7 +5,37 @@ import Input from '../index';
 export default [
   {
     examples: [
-      <Input type="text" placeholder="This is a placeholder" />,
+      <Input
+        type="text"
+        placeholder="This is a placeholder"
+      />,
+    ],
+  },
+  {
+    examples: [
+      <Input
+        type="number"
+        defaultValue="42"
+        onChange={
+          function(event) {
+            console.log(event.target.value); // eslint-disable-line no-console
+          }
+        }
+      />,
+    ],
+  },
+  {
+    examples: [
+      <Input
+        type="password"
+        label="Password"
+        testSection="input"
+      />,
+    ],
+  },
+  {
+    examples: [
+      <Input type="date" />,
     ],
   },
 ];